+++ apps/afisha/admin/performance.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from apps.afisha.models.masterclass import MasterClass
+from apps.afisha.models.master_class import MasterClass
 from apps.afisha.models.performance import Performance
 from apps.afisha.models.performance_media_review import PerformanceImage, PerformanceMediaReview, PerformanceReview
 from apps.afisha.models.reading import Reading
