+++ apps/afisha/models/master_class.py
