+++ apps/library/models/team_member.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.db.models import Q
 
-from apps.afisha.models.masterclass import MasterClass
+from apps.afisha.models.master_class import MasterClass
 from apps.afisha.models.performance import Performance
 from apps.afisha.models.reading import Reading
 from apps.core.models import BaseModel, Person
