+++ apps/library/views/search_result.py
